fix(dashboard/user-topup-modal): prevent clicking payment button multiple times (#436)
Pressing the "PAY" button starts the confirmation of the Payment Intent of the
current top-up. This sends a request to Stripe which will (if successful) return
a payment redirect link for iDEAL transactions.

Unfortunately, due to various reasons it is possible to accidentally press the
"PAY" button twice. This can be due to user error, device issues, or network
issues. The result is that sometimes, the request to confirm the Payment Intent
is send twice. This means that two payment redirect links are generated by
Stripe, but only the first one will be used.

Because Stripe has generated the second link (or even more) only that link is
still active for Stripe. In other words, successful payments through the first
link are not actively confirmed by Stripe. This means that that money is
"temporarily" considered lost and the top-up is not successful.

The money will be returned to the user by Stripe within 5 business days,
however, we should do everything in our power to ensure that this does not
happen (especially because people sometimes need to do large top-ups).

This change ensures that the "PAY" button is disabled after pressing it until
the user is redirected to the iDEAL payment page. I would have loved to use an
`Idempotency-Key` instead, however, it appears that this is not directly
supported by the `confirmPayment` method.